// src/_staart/interfaces/enum.d.ts
export declare enum Tokens {
    LOGIN = "auth",
    API_KEY = "api-key",
    ACCESS_TOKEN = "access-token",
    TWO_FACTOR = "2fa",
    REFRESH = "refresh",
    PASSWORD_RESET = "password-reset",
    EMAIL_VERIFY = "email-verify",
    EMAIL_RESEND = "email-resend",
    APPROVE_LOCATION = "approve-location",
    COUPON = "coupon",
    EMAIL_UPDATE = "email-update",
    CONFIRM_APPOINTMENT = "appointment"
}